Removed cs_connector and other unnecessary files, refactored (moved) search_helper.rb into csapi.rb
