--- example/files/cred.txt
@@ -1,1 +1,0 @@
-username;password
